+++ packages/spacecat-shared-gpt-client/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [@adobe/spacecat-shared-gpt-client-v1.6.21](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-gpt-client-v1.6.20...@adobe/spacecat-shared-gpt-client-v1.6.21) (2026-04-06)
+
+### Bug Fixes
+
+* **deps:** update external major (major) ([#1087](https://github.com/adobe/spacecat-shared/issues/1087)) ([72e1ab6](https://github.com/adobe/spacecat-shared/commit/72e1ab65892120f94ba409d5ab10370947329188))
+
 ## [@adobe/spacecat-shared-gpt-client-v1.6.20](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-gpt-client-v1.6.19...@adobe/spacecat-shared-gpt-client-v1.6.20) (2026-03-21)
 
 ### Bug Fixes